admin_schools_induction_coordinator route
